backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
SQLAlchemy==2.0.23
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic-settings==2.4.0
email-validator==2.2.0
numpy==1.26.4
opencv-python==4.10.0.84
insightface==0.7.3
onnxruntime==1.16.3

